Refactor: Removed extra files
